+++ src/dog_marker/api/v1/endpoints/user_entries.py
@@ -4,6 +4,7 @@ from uuid import UUID
 from fastapi import APIRouter, Depends, Response
 
 from dog_marker.dtypes.pagination import Pagination
+from dog_marker.database.schemas import warning_levels
 from .dependecies import get_service, query_pagination, authenticate_app
 from ..schemas import EntrySchema, CreateEntrySchema, UpdateEntrySchema
 from ..services import EntryService
@@ -14,10 +15,15 @@ router = APIRouter()
 @router.get("/{user_id}/entries", response_model=list[EntrySchema], operation_id="get_user_entries")
 async def get_user_entries(
     user_id: UUID,
+    warning_level: warning_levels = "information",
     page_info: Pagination = Depends(query_pagination),
     entry_service: EntryService = Depends(get_service(EntryService)),
 ) -> Iterable[EntrySchema]:
-    new_entry = entry_service.all(page_info=page_info, owner_id=user_id)
+    new_entry = entry_service.all(
+        page_info=page_info,
+        owner_id=user_id,
+        warning_level=warning_level,
+    )
     return new_entry