+++ src/dog_marker/api/v1/services/entry_service.py
@@ -9,6 +9,7 @@ from dog_marker.dtypes.coordinate import Coordinate
 from dog_marker.dtypes.pagination import Pagination
 from ..errors import NotAuthorizedError
 from ..schemas import EntrySchema, CreateEntrySchema, UpdateEntrySchema
+from dog_marker.database.schemas import WarningLevel, warning_levels
 
 
 class EntryService:
@@ -19,7 +20,7 @@ class EntryService:
         entry = self.entry_crud.get(entry_id)
         is_owner = entry.user_id == owner_id if owner_id else False
 
-        api_entry = EntrySchema.from_entry(entry, is_owner=is_owner)
+        api_entry = EntrySchema.from_db(entry, is_owner=is_owner)
 
         return api_entry
 
@@ -30,6 +31,7 @@ class EntryService:
         owner_id: UUID | None = None,
         coordinate: Coordinate | None = None,
         date_from: datetime.datetime | None = None,
+        warning_level: WarningLevel | warning_levels | None = None,
     ) -> Iterable[EntrySchema]:
         entries = self.entry_crud.all(
             user_id=user_id,
@@ -37,6 +39,7 @@ class EntryService:
             coordinate=coordinate,
             page_info=page_info,
             date_from=date_from,
+            warning_level=warning_level,
         )
         for entry in entries:
             is_owner = False
@@ -45,7 +48,7 @@ class EntryService:
             elif user_id is not None:
                 is_owner = entry.user_id == user_id
 
-            api_entry = EntrySchema.from_entry(entry, is_owner=is_owner)
+            api_entry = EntrySchema.from_db(entry, is_owner=is_owner)
 
             yield api_entry
 
@@ -57,14 +60,14 @@ class EntryService:
 
         entry = self.entry_crud.update(entry_id, update_entry)
 
-        api_entry = EntrySchema.from_entry(entry)
+        api_entry = EntrySchema.from_db(entry)
         api_entry.is_owner = old_entry.user_id == user_id
         return api_entry
 
     def create(self, user_id: UUID, data: CreateEntrySchema) -> EntrySchema:
         entry = self.entry_crud.create(user_id, data)
 
-        api_entry = EntrySchema.from_entry(entry)
+        api_entry = EntrySchema.from_db(entry)
         api_entry.is_owner = True
         return api_entry