+++ src/dog_marker/database/schemas/entry.py
@@ -5,6 +5,8 @@ from pydantic import BaseModel
 
 from dog_marker.dtypes.coordinate import Longitude, Latitude
 
+from .warning_level import WarningLevel
+
 
 class Entry(BaseModel):
     id: UUID
@@ -15,5 +17,6 @@ class Entry(BaseModel):
     image_delete_url: str | None
     longitude: Longitude
     latitude: Latitude
+    warning_level: WarningLevel = WarningLevel.information
     create_date: datetime
     update_date: datetime