+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## Develop
 
+- Add: WarningLevels
+
 ## v0.2.1
 
 - Fix: image_delete_url for non owner-users in get_all_entries